+++ lib/customized_roundbuttons.dart
@@ -1,0 +1,23 @@
+import 'package:flutter/material.dart';
+
+class RoundIconButton extends StatelessWidget {
+  IconData icon;
+  VoidCallback onPress;
+
+  RoundIconButton({required this.icon, required this.onPress});
+
+  @override
+  Widget build(BuildContext context) {
+    return RawMaterialButton(
+      onPressed: onPress,
+      elevation: 0.0,
+      constraints: const BoxConstraints.tightFor(
+        width: 56.0,
+        height: 56.0,
+      ),
+      shape: const CircleBorder(),
+      fillColor: const Color(0xFF4C4F5E),
+      child: Icon(icon),
+    );
+  }
+}
